requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
python-dotenv==1.2.2
requests==2.33.1
pydantic==2.13.3
urllib3==2.6.3

# DB
sqlalchemy==2.0.30
psycopg2-binary==2.9.9